# redwall-finance-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Redwall Finance
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /account/{accountNumber}:
    x-swagger-router-controller: account
    get:
      operationId: getAccount
      description: get an account
      parameters:
        - name: accountNumber
          description: account number
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AccountResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: updateAccount
      description: update an account
      parameters:
        - name: accountProperties
          description: Account properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/AccountPatchProperties"
        - name: accountNumber
          description: account number
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AccountResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /account/{accountNumber}/atm:
    x-swagger-router-controller: account
    post:
      operationId: doATMTransaction
      description: do an atm transaction
      parameters:
        - name: accountNumber
          description: account number
          type: integer
          in: path
          required: true
        - name: transactionProperties
          description: Transaction properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/ATMTransactionProperties"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TransactionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /account/{accountNumber}/debit:
    x-swagger-router-controller: account
    post:
      operationId: doDebitTransaction
      description: do a debit transaction
      parameters:
        - name: accountNumber
          description: account number
          type: integer
          in: path
          required: true
        - name: transactionProperties
          description: Transaction properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/DebitTransactionProperties"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TransactionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /account/{accountNumber}/transactions:
    x-swagger-router-controller: account
    get:
      operationId: getAccountTransactions
      description: get the transactions for an account
      parameters:
        - name: accountNumber
          description: account number
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTransactionsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /accounts:
    x-swagger-router-controller: accounts
    get:
      operationId: getAccounts
      description: get the accounts list
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetAccountsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /transaction/{transactionId}:
    x-swagger-router-controller: transaction
    get:
      operationId: getTransaction
      description: get an transaction
      parameters:
        - name: transactionId
          description: transaction number
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TransactionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /transactions:
    x-swagger-router-controller: transactions
    get:
      operationId: getTransactions
      description: get the transaction list
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTransactionsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user:
    x-swagger-router-controller: user
    post:
      operationId: addUser
      description: add a new user
      parameters:
        - name: userProperties
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserPostProperties"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{userId}:
    x-swagger-router-controller: user
    delete:
      operationId: removeUser
      description: remove a user
      parameters:
        - name: userId
          description: user ID
          type: integer
          in: path
          required: true
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: getUser
      description: get a user
      parameters:
        - name: userId
          description: user ID
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: updateUser
      description: update a user
      parameters:
        - name: userProperties
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserPatchProperties"
        - name: userId
          description: user ID
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{userId}/account:
    x-swagger-router-controller: user
    post:
      operationId: addUserAccount
      description: add a new account for this user
      parameters:
        - name: accountProperties
          description: Account properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/AccountPostProperties"
        - name: userId
          description: user ID
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AccountResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{userId}/accounts:
    x-swagger-router-controller: user
    get:
      operationId: getUserAccounts
      description: get the accounts belonging to a user
      parameters:
        - name: userId
          description: user ID
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetAccountsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    x-swagger-router-controller: users
    get:
      operationId: getUsers
      description: get the users list
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  AccountPatchProperties:
    properties:
      routingNumber:
        type: integer
      status:
        type: string
        enum: [OPEN, CLOSED, LOCKED]
  AccountPostProperties:
    required:
      - routingNumber
      - type
    properties:
      routingNumber:
        type: integer
      type:
        type: string
        enum: [CHECKING, SAVINGS]
  AccountResponse:
    required:
      - accountNumber
      - routingNumber
      - balanceInCents
      - type
      - status
      - createdAt
      - updatedAt
      - userId
    properties:
      accountNumber:
        type: integer
      routingNumber:
        type: integer
      balanceInCents:
        type: integer
      type:
        type: string
        enum: [CHECKING, SAVINGS]
      status:
        type: string
        enum: [OPEN, CLOSED, LOCKED]
      createdAt:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
  ATMTransactionProperties:
    required:
      - amountInCents
      - type
    properties:
      amountInCents:
        type: integer
      type:
        type: string
        enum: [DEPOSIT, WITHDRAWAL]
  DebitTransactionProperties:
    required:
      - amountInCents
      - type
    properties:
      amountInCents:
        type: integer
      type:
        type: string
        enum: [CHARGE, CREDIT]
  GetAccountsListResponse:
    type: array
    items:
      $ref: "#/definitions/AccountResponse"
  GetTransactionsListResponse:
    type: array
    items:
      $ref: "#/definitions/TransactionResponse"
  GetUsersListResponse:
    type: array
    items:
      $ref: "#/definitions/UserResponse"
  TransactionResponse:
    required:
      - transactionId
      - amountInCents
      - description
      - fromAccountType
      - fromAccountNumber
      - fromRoutingNumber
      - toAccountType
      - toAccountNumber
      - toRoutingNumber
      - createdAt
      - updatedAt
    properties:
      transactionId:
        type: integer
      amountInCents:
        type: integer
      description:
        type: string
      fromAccountType:
        type: string
        enum: ['ATM', 'DEBIT', 'EXTERNAL', 'INTERNAL']
      fromAccountNumber:
        type: integer
      fromRoutingNumber:
        type: integer
      toAccountType:
        type: string
        enum: ['ATM', 'DEBIT', 'EXTERNAL', 'INTERNAL']
      toAccountNumber:
        type: integer
      toRoutingNumber:
        type: integer
      createdAt:
        type: string
      updatedAt:
        type: string
  UserPatchProperties:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
  UserPostProperties:
    required:
      - firstName
      - lastName
      - email
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
  UserResponse:
    required:
      - userId
      - firstName
      - lastName
      - email
      - createdAt
      - updatedAt
    properties:
      userId:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string